Add lst_to_array test for empty lst.

+++ tests/lst/test_ft_lstadd_back.c
@@ -32,7 +32,7 @@ Test(ft_lstadd_back)
 	ft_lstadd_back(&lst, new);
 
 	if (NOT has_been_unstable)
-		massert(ft_lstsize(lst), 2);
+		massert(ft_lstsize(lst), (size_t)2);
 
 	ft_lstclear(&lst, NULL);
 }
